+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ### Removed
 
 - character parsers that were aliases to their `*1` version: eol, alpha, digit, hex_digit, oct_digit, alphanumeric, space, multispace
+- `count_fixed` macro
 
 ## 4.2.3 - 2019-03-23